+++ lib/trellis/plugins/callback/vars.py
@@ -5,10 +5,11 @@ import re
 import sys
 
 from __main__ import cli
-from ansible.compat.six import iteritems
+from ansible.module_utils.six import iteritems
 from ansible.errors import AnsibleError
 from ansible.parsing.dataloader import DataLoader
 from ansible.parsing.yaml.objects import AnsibleMapping, AnsibleSequence, AnsibleUnicode
+from ansible.playbook.play_context import PlayContext
 from ansible.plugins.callback import CallbackBase
 from ansible.template import Templar
 
@@ -58,7 +59,6 @@ class CallbackModule(CallbackBase):
 
         strings = {
             '--connection': 'connection',
-            '--inventory-file': 'inventory',
             '--private-key': 'private_key_file',
             '--ssh-common-args': 'ssh_common_args',
             '--ssh-extra-args': 'ssh_extra_args',
@@ -69,6 +69,9 @@ class CallbackModule(CallbackBase):
         for option,value in strings.iteritems():
             if getattr(self._options, value, False):
                 options.append("{0}='{1}'".format(option, str(getattr(self._options, value))))
+
+        for inventory in getattr(self._options, 'inventory'):
+            options.append("--inventory='{}'".format(str(inventory)))
 
         if getattr(self._options, 'ask_vault_pass', False):
             options.append('--ask-vault-pass')
@@ -87,8 +90,9 @@ class CallbackModule(CallbackBase):
 
     def v2_playbook_on_play_start(self, play):
         for host in play.get_variable_manager()._inventory.list_hosts(play.hosts[0]):
-            hostvars = play.get_variable_manager().get_vars(loader=self.loader, play=play, host=host)
+            hostvars = play.get_variable_manager().get_vars(play=play, host=host)
             self.raw_vars(play, host, hostvars)
+            host.vars['ssh_args_default'] = PlayContext(play=play, options=self._options)._ssh_args.default
             host.vars['cli_options'] = self.cli_options()
             host.vars['cli_ask_pass'] = getattr(self._options, 'ask_pass', False)
             host.vars['cli_ask_become_pass'] = getattr(self._options, 'become_ask_pass', False)